% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDFE.R
\name{erfc}
\alias{erfc}
\title{erfc}
\usage{
erfc(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A numeric vector, i.e., the erfc values.
}
\description{
This is the complementary error function.
}

